Player: Interaction & UIManager states
Made Camera a prefab, added EventSystem to Object.
Added StartGameButton to GameStartScreen in order to link it with the GameManager on Start.
Added Player to GameManager as a basic input handling class. Added CheckInput and PressedPrimaryButton/PressedSecondaryButton.
Added Empty ToggleCellState() and CrossCell to LevelManager.
Added SwitchState(GameState) to UIManager.
Renamed some buttons in ProjectSettings->Input